Experimenting with MongoDB/Mongosh
