+++ back-end/index.js
@@ -1,15 +1,25 @@
 const express = require("express");
 const mongoose  = require("mongoose");
 const app = express();
-const {Location, User} = require("./seed.js")
+const {Location, User} = require("./seed.js");
+const {connectToDb, getDb} = require("./db.js");
 
-app.listen(9090, () => {
-    console.log("Listening on 9090")
+let db
+
+connectToDb((err)=>{
+    if (!err) {
+        app.listen(9090, () => {
+            console.log("Listening on 9090")
+        })
+        db = getDb()
+    }
 })
 
-mongoose.connect("mongodb+srv://jsmilezz052:BnkJNB4pGloVf1o4@cluster0.iexwylq.mongodb.net/sample_mflix?retryWrites=true&w=majority&appName=Cluster0").then(()=>{
-    console.log("Connected to DB")
-}).catch((err)=>{
-    console.log("Connection failed", err)
-})
+
+// mongoose.connect("mongodb+srv://jsmilezz052:BnkJNB4pGloVf1o4@cluster0.iexwylq.mongodb.net/DayPlanner_DB?retryWrites=true&w=majority&appName=Cluster0").then(()=>{
+    
+//     console.log("Connected to DB")
+// }).catch((err)=>{
+//     console.log("Connection failed", err)
+// })
